# perf: remove spread opts and toString of integrity