--- a/lib/index.js
+++ b/lib/index.js
@@ -210,6 +210,41 @@ class Hash {
   }
 }
 
+function integrityHashToString(toString, sep, opts, hashes) {
+  if (!hashes || hashes.length === 0)
+    return toString;
+
+  const toStringIsNotEmpty = toString !== '';
+
+  let shouldAddFirstSep = false;
+  let complement = '';
+
+  const lastIndex = hashes.length - 1
+
+  for (let i = 0; i < lastIndex; i++) {
+    const hashString = Hash.prototype.toString.call(hashes[i], opts)
+
+    if (hashString) {
+      shouldAddFirstSep = true;
+
+      complement += hashString
+      complement += sep
+    }
+  }
+
+  const finalHashString = Hash.prototype.toString.call(hashes[lastIndex], opts)
+
+  if (finalHashString) {
+    shouldAddFirstSep = true;
+    complement += finalHashString
+  }
+
+  if (toStringIsNotEmpty && shouldAddFirstSep)
+    return toString + sep + complement;
+
+  return toString + complement 
+}
+
 class Integrity {
   get isIntegrity () {
     return true
@@ -225,15 +260,30 @@ class Integrity {
 
   toString (opts) {
     let sep = opts && opts.sep || ' '
-    if (opts.strict) {
+    let toString = '';
+
+    if (opts && opts.strict) {
       // Entries must be separated by whitespace, according to spec.
       sep = sep.replace(/\S+/g, ' ')
+
+      if (this.sha512 && this.sha512.length > 0) {
+        toString = integrityHashToString(toString, sep,opts, this['sha512'])
+      }
+
+      if (this.sha384 && this.sha384.length > 0) {
+        toString = integrityHashToString(toString, sep,opts, this['sha384'])
+      }
+
+      if (this.sha256 && this.sha256.length > 0) {
+        toString = integrityHashToString(toString,sep, opts, this['sha256'])
+      }
+    } else {
+      for (const hash of Object.keys(this)) {
+        toString = integrityHashToString(toString, sep,opts, this[hash])
+      }
     }
-    return Object.keys(this).map(k => {
-      return this[k].map(hash => {
-        return Hash.prototype.toString.call(hash, opts)
-      }).filter(x => x.length).join(sep)
-    }).filter(x => x.length).join(sep)
+
+    return toString;
   }
 
   concat (integrity, opts) {