--- a/lib/index.js
+++ b/lib/index.js
@@ -40,33 +40,41 @@ class IntegrityStream extends MiniPass {
   constructor (opts) {
     super()
     this.size = 0
-    this.opts = opts
+    this.opts = opts || Object.create(null)
 
     // may be overridden later, but set now for class consistency
     this[_getOptions]()
 
     // options used for calculating stream.  can't be changed.
-    const algorithms = opts && opts.algorithms || defaultOpts.algorithms
-    this.algorithms = Array.from(
-      new Set(algorithms.concat(this.algorithm ? [this.algorithm] : []))
-    )
+    if (opts && opts.algorithms) {
+      const algorithms = opts.algorithms
+      this.algorithms = Array.from(
+        new Set(algorithms.concat(this.algorithm ? [this.algorithm] : []))
+      )
+    } else {
+      this.algorithms = defaultOpts.algorithms
+
+      if (this.algorithm !== null && this.algorithm !== defaultOpts.algorithms[0])
+        this.algorithms.push(this.algorithm);
+    }
+
     this.hashes = this.algorithms.map(crypto.createHash)
   }
 
   [_getOptions] () {
-    const {
-      integrity,
-      size,
-      options,
-    } = { ...defaultOpts, ...this.opts }
-
     // For verification
-    this.sri = integrity ? parse(integrity, this.opts) : null
-    this.expectedSize = size
-    this.goodSri = this.sri ? !!Object.keys(this.sri).length : false
-    this.algorithm = this.goodSri ? this.sri.pickAlgorithm(this.opts) : null
+    this.sri = this.opts.integrity ? parse(this.opts.integrity, this.opts) : null
+    this.expectedSize = this.opts.size
+    this.goodSri = this.sri instanceof Integrity 
+      ? !!Object.keys(this.sri).length 
+      : this.sri instanceof Hash
+    this.algorithm = this.goodSri 
+      ? this.sri instanceof Integrity
+        ? this.sri.pickAlgorithm(this.opts)
+        : this.sri.algorithm
+      : null
     this.digests = this.goodSri ? this.sri[this.algorithm] : null
-    this.optString = getOptString(options)
+    this.optString = getOptString(this.opts.options)
   }
 
   on (ev, handler) {
@@ -180,6 +188,17 @@ class Hash {
 
   toJSON () {
     return this.toString()
+  }
+
+  match (integrity, opts) {
+    const other = parse(integrity, opts)
+    if (!other) {
+      return false
+    }
+    const algo = other instanceof Integrity
+      ? other.pickAlgorithm(opts)
+      : other
+    return algo.digest === this.digest ? algo : false
   }
 
   toString (opts) {
@@ -433,7 +452,7 @@ function fromStream (stream, opts) {
       sri = s
     })
     istream.on('end', () => resolve(sri))
-    istream.on('data', () => {})
+    istream.resume()
   })
 }
 
@@ -500,7 +519,7 @@ function checkStream (stream, sri, opts) {
       verified = s
     })
     checker.on('end', () => resolve(verified))
-    checker.on('data', () => {})
+    checker.resume()
   })
 }