--- a/lib/index.js
+++ b/lib/index.js
@@ -46,7 +46,7 @@ class IntegrityStream extends MiniPass {
     this[_getOptions]()
 
     // options used for calculating stream.  can't be changed.
-    if (opts && opts.algorithms) {
+    if (opts?.algorithms) {
       const algorithms = opts.algorithms
       this.algorithms = Array.from(
         new Set(algorithms.concat(this.algorithm ? [this.algorithm] : []))
@@ -151,7 +151,7 @@ class Hash {
   }
 
   constructor (hash, opts) {
-    const strict = opts && opts.strict
+    const strict = opts?.strict
     this.source = hash.trim()
 
     // set default values so that we make V8 happy to
@@ -202,7 +202,7 @@ class Hash {
   }
 
   toString (opts) {
-    if (opts && opts.strict) {
+    if (opts?.strict) {
       // Strict mode enforces the standard as close to the foot of the
       // letter as it can.
       if (!(
@@ -278,10 +278,10 @@ class Integrity {
   }
 
   toString (opts) {
-    let sep = opts && opts.sep || ' '
+    let sep = opts?.sep || ' '
     let toString = '';
 
-    if (opts && opts.strict) {
+    if (opts?.strict) {
       // Entries must be separated by whitespace, according to spec.
       sep = sep.replace(/\S+/g, ' ')
 
@@ -351,7 +351,7 @@ class Integrity {
   }
 
   pickAlgorithm (opts) {
-    const pickAlgorithm = opts && opts.pickAlgorithm || defaultOpts.pickAlgorithm;
+    const pickAlgorithm = opts?.pickAlgorithm || defaultOpts.pickAlgorithm;
     const keys = Object.keys(this)
     return keys.reduce((acc, algo) => {
       return pickAlgorithm(acc, algo) || acc
@@ -378,7 +378,7 @@ function parse (sri, opts) {
 function _parse (integrity, opts) {
   // 3.4.3. Parse metadata
   // https://w3c.github.io/webappsec-subresource-integrity/#parse-metadata
-  if (opts && opts.single) {
+  if (opts?.single) {
     return new Hash(integrity, opts)
   }
   const hashes = integrity.trim().split(/\s+/).reduce((acc, string) => {
@@ -408,7 +408,7 @@ function stringify (obj, opts) {
 
 module.exports.fromHex = fromHex
 function fromHex (hexDigest, algorithm, opts) {
-  const optString = getOptString(opts && opts.options)
+  const optString = getOptString(opts?.options)
   return parse(
     `${algorithm}-${
       Buffer.from(hexDigest, 'hex').toString('base64')
@@ -418,8 +418,8 @@ function fromHex (hexDigest, algorithm, opts) {
 
 module.exports.fromData = fromData
 function fromData (data, opts) {
-  const algorithms = opts && opts.algorithms || defaultOpts.algorithms
-  const optString = getOptString(opts && opts.options)
+  const algorithms = opts?.algorithms || defaultOpts.algorithms
+  const optString = getOptString(opts?.options)
   return algorithms.reduce((acc, algo) => {
     const digest = crypto.createHash(algo).update(data).digest('base64')
     const hash = new Hash(
@@ -460,7 +460,7 @@ module.exports.checkData = checkData
 function checkData (data, sri, opts) {
   sri = parse(sri, opts)
   if (!sri || !Object.keys(sri).length) {
-    if (opts && opts.error) {
+    if (opts?.error) {
       throw Object.assign(
         new Error('No valid integrity hashes to check against'), {
           code: 'EINTEGRITY',
@@ -475,7 +475,7 @@ function checkData (data, sri, opts) {
   const newSri = parse({ algorithm, digest })
   const match = newSri.match(sri, opts)
   opts = opts || Object.create(null)
-  if (match || !(opts && opts.error)) {
+  if (match || !(opts?.error)) {
     return match
   } else if (typeof opts.size === 'number' && (data.length !== opts.size)) {
     /* eslint-disable-next-line max-len */
@@ -530,8 +530,8 @@ function integrityStream (opts = Object.create(null)) {
 
 module.exports.create = createIntegrity
 function createIntegrity (opts) {
-  const algorithms = opts && opts.algorithms || defaultOpts.algorithms
-  const optString = getOptString(opts && opts.options)
+  const algorithms = opts?.algorithms || defaultOpts.algorithms
+  const optString = getOptString(opts?.options)
 
   const hashes = algorithms.map(crypto.createHash)