@wraithgar 2023-04-01
```sh
$ npm view ssri engines
{ node: '^14.17.0 || ^16.13.0 || >=18.0.0' }
```

we can use opts?.algorithms et al