@wraithgar 2023-04-01
I really appreciate the time you put into this, can we maybe break this up so that the changes aren't so huge in one PR? Specifically I'm worried about the default opts handling.  It'd be nice to isolate those changes against the other tweaks.