@H4ad 2023-04-01
@wraithgar What about 3 PRs:

- One for options
- One for faster to string
- One for stream

It will be better?