@H4ad 2023-04-01
@wraithgar First one created at https://github.com/npm/ssri/pull/72, when it was merged, I create the next one.